// typechain/interfaces/IV1FeedRegistry.ts
export type RequestHeaderPatchStruct = { key: string; ciphertext: BytesLike };

export type RequestHeaderPatchStructOutput = [
  key: string,
  ciphertext: string
] & { key: string; ciphertext: string };

export type QueryParameterPatchStruct = { key: string; ciphertext: BytesLike };

export type QueryParameterPatchStructOutput = [
  key: string,
  ciphertext: string
] & { key: string; ciphertext: string };

export type HTTPPrivatePatchStruct = {
  pathSuffix: BytesLike;
  headers: RequestHeaderPatchStruct[];
  parameters: QueryParameterPatchStruct[];
  body: BytesLike;
};

export type HTTPPrivatePatchStructOutput = [
  pathSuffix: string,
  headers: RequestHeaderPatchStructOutput[],
  parameters: QueryParameterPatchStructOutput[],
  body: string
] & {
  pathSuffix: string;
  headers: RequestHeaderPatchStructOutput[];
  parameters: QueryParameterPatchStructOutput[];
  body: string;
};

export type RequestHeaderStruct = { key: string; value: string };

export type RequestHeaderStructOutput = [key: string, value: string] & {
  key: string;
  value: string;
};

export type QueryParameterStruct = { key: string; value: string };

export type QueryParameterStructOutput = [key: string, value: string] & {
  key: string;
  value: string;
};

export type HTTPRequestStruct = {
  method: BigNumberish;
  host: string;
  path: string;
  headers: RequestHeaderStruct[];
  parameters: QueryParameterStruct[];
  body: BytesLike;
};

export type HTTPRequestStructOutput = [
  method: bigint,
  host: string,
  path: string,
  headers: RequestHeaderStructOutput[],
  parameters: QueryParameterStructOutput[],
  body: string
] & {
  method: bigint;
  host: string;
  path: string;
  headers: RequestHeaderStructOutput[];
  parameters: QueryParameterStructOutput[];
  body: string;
};

export type FeedStruct = {
  request: HTTPRequestStruct;
  patch: HTTPPrivatePatchStruct;
  schema: string;
  filter: string;
};

export type FeedStructOutput = [
  request: HTTPRequestStructOutput,
  patch: HTTPPrivatePatchStructOutput,
  schema: string,
  filter: string
] & {
  request: HTTPRequestStructOutput;
  patch: HTTPPrivatePatchStructOutput;
  schema: string;
  filter: string;
};